fix feature delete for superadmin

+++ Views/Barang/barang.php
@@ -72,7 +72,7 @@ require_once __DIR__ . '/../Layout/header.php';
                 Export EXCEL
             </a>
 
-            <a href="barang-created.php" 
+            <a href="barang-created.php"
                 class="bg-blue-600 text-white px-4 py-2 rounded-lg hover:bg-blue-700 flex items-center gap-2">
 
                 <!-- Icon Plus -->
@@ -185,12 +185,13 @@ require_once __DIR__ . '/../Layout/header.php';
                                 class="bg-yellow-500 text-white px-3 py-1 rounded">
                                 Edit
                             </a>
-
-                            <a href="barang-deleted.php?id=<?= $row['id'] ?>"
-                                onclick="return confirm('Yakin hapus Barang?')"
-                                class="px-4 py-2 bg-red-600 text-white rounded-lg hover:bg-red-700">
-                                Hapus
-                            </a>
+                            <?php if (isset($_SESSION['role']) && $_SESSION['role'] === 'superadmin'): ?>
+                                <a href="barang-deleted.php?id=<?= $row['id'] ?>"
+                                    onclick="return confirm('Yakin hapus Barang?')"
+                                    class="px-4 py-2 bg-red-600 text-white rounded-lg hover:bg-red-700">
+                                    Hapus
+                                </a>
+                            <?php endif; ?>
 
                         </td>
